novas verificações

+++ webservice.php
@@ -1,8 +1,11 @@
 <?php
+/***************************IMPORTS***************************/
 require_once 'lib/nusoap.php';
 require_once 'bd.php';
 require_once 'xml.php';
+/***************************IMPORTS***************************/
 
+/***************************REGISTROS***************************/
 $server = new soap_server();
 $server->configureWSDL("webservice", "urn:webservice");
 
@@ -38,21 +41,32 @@ $server->register("DeletaUsuario",
 
 $HTTP_RAW_POST_DATA = isset($HTTP_RAW_POST_DATA) ? $HTTP_RAW_POST_DATA : '';
 $server->service($HTTP_RAW_POST_DATA);
+/***************************REGISTROS***************************/
 
+/***************************FUNÇÕES***************************/
 function CadastraUsuario($nome, $email, $senha){
+    
+    if(!isset($nome) || !isset($email) || !isset($senha)){
+        $xml = xml('ERRO', 'Falha na Insercao', 'Um dos parametros em branco');
+        return $xml;
+    }
     
     $stmt = $GLOBALS['db']->prepare(
     'INSERT INTO usuarios (nome, email, senha)
     VALUES (:nome, :email, :senha)');
-    $stmt->execute(array(
+    $status = $stmt->execute(array(
         'nome' => $nome,
         'email' => $email,
         'senha' => $senha
     ));
     
-    $id = $GLOBALS['db']->lastInsertId();
+    if ($status){
+        $id = $GLOBALS['db']->lastInsertId();
     
-    $xml = xml('OK', 'Insercao feita com sucesso', 'Usuario com id '.$id.' criado com sucesso');
+        $xml = xml('OK', 'Insercao feita com sucesso', 'Usuario com id '.$id.' criado com sucesso');
+    } else {
+        $xml = xml('ERRO', 'Falha na Insercao', $stmt->errorInfo()); 
+    }
     
     return $xml;
     
@@ -60,21 +74,30 @@ function CadastraUsuario($nome, $email, $senha){
 
 function ValidaSecao($email, $senha){
     
+    if(!isset($email) || !isset($senha)){
+        $xml = xml('ERRO', 'Falha na Consulta', 'Um dos parametros em branco');
+        return $xml;
+    }
+
     $stmt = $GLOBALS['db']->prepare(
     'SELECT * FROM usuarios
     WHERE email = :email AND senha = :senha');
-    $stmt->execute(array(
+    $status = $stmt->execute(array(
         'email' => $email,
         'senha' => $senha
     ));
     $resultado = $stmt->fetch(PDO::FETCH_ASSOC);
 
-    if(!empty($resultado)){
-        $resultado[validade] = 1;
-        $xml = xml('OK', 'Consulta feita com sucesso', $resultado);
+    if ($status){
+        if(!empty($resultado)){
+            $resultado[validade] = 1;
+            $xml = xml('OK', 'Consulta feita com sucesso', $resultado);
+        } else {
+            $resultado[validade] = -1;
+            $xml = xml('OK', 'Consulta não retornou nenhum valor', $resultado);
+        }
     } else {
-        $resultado[validade] = -1;
-        $xml = xml('BAD', 'Consulta não retornou nenhum valor', $resultado);
+        $xml = xml('ERRO', 'Falha na Consulta', $stmt->errorInfo()); 
     }
     
     return $xml;
@@ -83,15 +106,25 @@ function ValidaSecao($email, $senha){
 
 function DeletaUsuario($id){
     
+    if(!isset($id)){
+        $xml = xml('ERRO', 'Falha na Insercao', 'ID em branco');
+        return $xml;
+    }
+    
     $stmt = $GLOBALS['db']->prepare(
     'DELETE FROM usuarios
     WHERE id = :id');
-    $stmt->execute(array(
+    $status = $stmt->execute(array(
         'id' => $id
     ));
     
-    $xml = xml('OK', 'Remocao feita com sucesso', 'Usuario com id '.$id.' deletado com sucesso');
+    if ($status){
+        $xml = xml('OK', 'Remocao feita com sucesso', 'Usuario com id '.$id.' deletado com sucesso');
+    } else {
+        $xml = xml('ERRO', 'Falha na Remocao', $stmt->errorInfo()); 
+    }
     
     return $xml;
     
 }
+/***************************FUNÇÕES***************************/
